requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
openai>=1.0.0
python-dotenv
